Ajout d'une verification du fichier envoyé sur le serveur

+++ tp_a_rendre/addImgServ.php
@@ -22,20 +22,24 @@ if (isset($_SESSION["email"]) and $_SESSION["email"] != NULL) {
 
         if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 
-            $imageActuelle = "images/profiles/".selectNomImageFromEmail($_SESSION['email'])['nomImage'];
+            $imageActuelle = "images/profiles/" . selectNomImageFromEmail($_SESSION['email'])['nomImage'];
 
-            if ( $imageActuelle != NULL && file_exists($imageActuelle)) {
-                unlink($imageActuelle);
-            }
+            $imageName = $_SESSION['email'] . date("Y-m-d") . time();
 
-            $imageName = $_SESSION['email'].date("Y-m-d").time();   
+            if (mime_content_type($_FILES['fileToUpload']['tmp_name']) == 'image/png' || mime_content_type($_FILES['fileToUpload']['tmp_name']) == 'image/jpeg' || mime_content_type($_FILES['fileToUpload']['tmp_name']) == 'image/gif') {
 
-            if ($_FILES['fileToUpload']['size'] < 500000) {
-                updateImageMembre($_SESSION['email'], $imageName);
-                move_uploaded_file($_FILES['fileToUpload']['tmp_name'], "images/profiles/$imageName");
-                header("Location: addImgServ.php");
+                if ($_FILES['fileToUpload']['size'] < 500000) {
+                    updateImageMembre($_SESSION['email'], $imageName);
+                    move_uploaded_file($_FILES['fileToUpload']['tmp_name'], "images/profiles/$imageName");
+                    if ($imageActuelle != NULL && file_exists($imageActuelle)) {
+                        unlink($imageActuelle);
+                    }
+                    header("Location: addImgServ.php");
+                } else {
+                    echo "<p>Fichier trop lourd, veuillez sélectionner une image de moins de 500ko</p>";
+                }
             } else {
-                echo "<p>Fichier trop lourd, veuillez sélectionner une image de moins de 500ko</p>";
+                echo "<p>Fichier invalide</p>";
             }
         }
         ?>
